update cluster template
