+++ templates/cluster-tsp/input.json
@@ -1,0 +1,111 @@
+{
+  "vehicles": ["v1", "v2"],
+  "stops": [
+    {
+      "id": "West",
+      "position": {
+        "lon": -74.04866,
+        "lat": 4.69018
+      }
+    },
+    {
+      "id": "North",
+      "position": {
+        "lon": -74.044215,
+        "lat": 4.693907
+      }
+    },
+    {
+      "id": "East",
+      "position": {
+        "lon": -74.039745,
+        "lat": 4.689886
+      }
+    },
+    {
+      "id": "South",
+      "position": {
+        "lon": -74.044219,
+        "lat": 4.686293
+      }
+    },
+    {
+      "id": "North East",
+      "position": {
+        "lon": -74.041242,
+        "lat": 4.692397
+      }
+    },
+    {
+      "id": "South West",
+      "position": {
+        "lon": -74.047572,
+        "lat": 4.687877
+      }
+    },
+    {
+      "id": "North West",
+      "position": {
+        "lon": -74.046988,
+        "lat": 4.692952
+      }
+    },
+    {
+      "id": "South East",
+      "position": {
+        "lon": -74.040685,
+        "lat": 4.687558
+      }
+    },
+    {
+      "id": "Place 1",
+      "position": {
+        "lon": -74.047,
+        "lat": 4.69014
+      }
+    },
+    {
+      "id": "Place 2",
+      "position": {
+        "lon": -74.0485,
+        "lat": 4.69017
+      }
+    },
+    {
+      "id": "Place 3",
+      "position": {
+        "lon": -74.049,
+        "lat": 4.69012
+      }
+    },
+    {
+      "id": "Place 4",
+      "position": {
+        "lon": -74.048,
+        "lat": 4.69015
+      }
+    }
+  ],
+  "starts": [
+    {
+      "lon": -74.04866,
+      "lat": 4.69018
+    },
+    {
+      "lon": -74.04866,
+      "lat": 4.69018
+    }
+  ],
+  "ends": [
+    {
+      "lon": -74.04866,
+      "lat": 4.69018
+    },
+    {
+      "lon": -74.04866,
+      "lat": 4.69018
+    }
+  ],
+  "quantities": [-1, -1, -1, -1, -1, -1, -1, -1, 0, 0, 0, 0],
+  "capacities": [4, 4]
+}
